blind-nav-fullstack-package/blind-nav-fullstack-package/blind-nav-backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
pillow
numpy
python-dotenv
